概率分布随机种子调整为int类型.

+++ src/main/java/com/jstarcraft/ai/math/algorithm/probability/QuantityProbability.java
@@ -26,14 +26,14 @@ public class QuantityProbability implements Probability<Double>, ModemCycle {
 
     private String randomClass;
 
-    private long randomSeed;
+    private int randomSeed;
 
     private transient RandomGenerator random;
 
     QuantityProbability() {
     }
 
-    public QuantityProbability(Class<? extends RandomGenerator> randomClazz, long randomSeed, Class<? extends AbstractRealDistribution> distributionClazz, Object... distributionParameters) {
+    public QuantityProbability(Class<? extends RandomGenerator> randomClazz, int randomSeed, Class<? extends AbstractRealDistribution> distributionClazz, Object... distributionParameters) {
         this.randomSeed = randomSeed;
         this.random = ReflectionUtility.getInstance(randomClazz, randomSeed);
         this.distributionParameters = distributionParameters;
